+++ srcDLL/DllMain.cpp
@@ -1,4 +1,4 @@
-#include "IpDropDown.h"
+#include "GameModules/IpDropDown.h"
 #include "IniModuleLoader.h"
 #include "StringConversion.h"
 #include "OP2Memory.h"
