+++ srcStatic/op2extStatic.vcxproj
@@ -133,7 +133,7 @@
     <ClCompile Include="FileSystemHelper.cpp" />
     <ClCompile Include="GlobalDefines.cpp" />
     <ClCompile Include="IniModuleLoader.cpp" />
-    <ClCompile Include="IpDropDown.cpp" />
+    <ClCompile Include="GameModules\IpDropDown.cpp" />
     <ClCompile Include="ConsoleModuleLoader.cpp" />
     <ClCompile Include="Logger.cpp" />
     <ClCompile Include="op2ext-Internal.cpp" />
@@ -149,7 +149,7 @@
     <ClInclude Include="GameModules\GameModule.h" />
     <ClInclude Include="GlobalDefines.h" />
     <ClInclude Include="IniModuleLoader.h" />
-    <ClInclude Include="IpDropDown.h" />
+    <ClInclude Include="GameModules\IpDropDown.h" />
     <ClInclude Include="ConsoleModuleLoader.h" />
     <ClInclude Include="Logger.h" />
     <ClInclude Include="op2ext-Internal.h" />
